+++ cmd/mistryd/testdata/projects/result-cache-exitcode/docker-entrypoint.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -e
-
-date +%S%N > artifacts/out.txt
-
-exit 33