+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Breaking changes are prefixed with a "[BREAKING]" label.
 
 ### Changed
 
+- Failed jobs are no longer cached [[#62](https://github.com/skroutz/mistry/pull/62)]
 - **[BREAKING]** client/server: Client and server binaries are renamed to "mistryd" and "mistry" respectively.
   Also project is now go-gettable. [[abbfb58](https://github.com/skroutz/mistry/commit/abbfb58d5a2aaf3eaebf9408d81ec7d459326416)]
 - client: default host is now 0.0.0.0